+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN6CQE8S84MYTWEBNPXBMCZW"
+        "01KN6MC8M6PH2EZW49K7DDP7D4"
     )
     r"""The unique identifier of the data source"""